move cli over to toolpad-app
