+++ packages/toolpad-app/cli.js
@@ -1,2 +1,5 @@
 #!/usr/bin/env node
-require('./dist/cli');
+
+const { default: cli } = require('./dist/cli');
+
+cli(process.argv.slice(2));
