+++ packages/toolpad/index.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+const { default: cli } = require('@mui/toolpad-app/dist/cli');
+
+cli(process.argv.slice(2));
