+++ packages/toolpad/tsconfig.cjs.json
@@ -3,6 +3,5 @@
   "compilerOptions": {
     "module": "CommonJS",
     "outDir": "dist/cjs"
-  },
-  "exclude": ["**/*.mts"]
+  }
 }
