+++ packages/toolpad/tsconfig.json
@@ -14,5 +14,5 @@
     "target": "ES2020",
     "skipLibCheck": true
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.mts"]
+  "include": ["src/**/*.ts", "src/**/*.tsx"]
 }
